Extract shared docContent component with code-block copy; adopt on desktop
