+++ client-app/src/core/docs/index.ts
@@ -1,0 +1,4 @@
+export {docContent} from './DocContent';
+export {DocViewModel} from './DocViewModel';
+export * from './types';
+export * from './DocUtils';
